refactor(obd2): extract elm327_commands + elm327_parsers (Refs #563)
Refs #563 — elm327_protocol.dart was 599 LOC, split into command
constants + response parsers. No behavior change; existing tests pass.

- elm327_commands.dart (250 LOC): AT setup commands, Mode 01/09 PID
  request strings, manufacturer odometer catalog, VehicleBrand enum
  and vehicleBrandFromVin helper.
- elm327_parsers.dart (349 LOC): pure string->value decoders for
  Mode 01, Mode 09 and Mode 22 responses.
- elm327_protocol.dart (157 LOC): backwards-compatible facade that
  re-exports both libraries and keeps Elm327Protocol.xxx statics
  working for the ~30 existing call sites (trip_recording_controller,
  obd2_service, onboarding_obd2_connector, tests).
